fix product service security

+++ src/main/java/com/project/zaiika/services/placeServices/ProductServiceImpl.java
@@ -4,9 +4,8 @@ import com.project.zaiika.models.placeModels.Ingredient;
 import com.project.zaiika.models.placeModels.Product;
 import com.project.zaiika.repositories.placesRepository.IngredientRepository;
 import com.project.zaiika.repositories.placesRepository.ProductRepository;
+import com.project.zaiika.services.util.ContextService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,10 +15,13 @@ import java.util.List;
 public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
     private final IngredientRepository ingredientRepository;
+    private final ContextService ctx;
 
 
     @Override
     public List<Product> getAllProductFromMenu(long menuId) {
+        checkPermission(menuId);
+
         var products = productRepository.findAllByMenuId(menuId);
 
         List<Long> productsId = products.stream().map(Product::getId).toList();
@@ -34,6 +36,8 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void addProductToMenu(Product product) {
+        checkPermission(product.getMenuId());
+
         var saveProduct = productRepository.save(product);
         for (Ingredient ingredient : product.getComposition()) {
             ingredient.setProductId(saveProduct.getId());
@@ -44,6 +48,8 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void updateProduct(Product updateProduct) {
+        checkPermission(updateProduct.getMenuId());
+
         validateProductId(updateProduct.getId());
 
         for (Ingredient ingredient : updateProduct.getComposition()) {
@@ -57,6 +63,7 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProductById(long menuId, long productId) {
+        checkPermission(menuId);
         validateProductId(productId);
 
         productRepository.deleteProductByMenuIdAndId(menuId, productId);
@@ -66,6 +73,13 @@ public class ProductServiceImpl implements ProductService {
     private void validateProductId(long id) {
         if (id <= 0) {
             throw new IllegalArgumentException("Invalid product id '" + id + "'");
+        }
+    }
+
+    private void checkPermission(long menuId) {
+        var menu = ctx.getMenu(menuId);
+        if (menu == null) {
+            throw new IllegalArgumentException("permission denied");
         }
     }
 }
